% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_palettes.R
\docType{data}
\name{BB_colors}
\alias{BB_colors}
\title{Custom palette for ggplot2 graphs}
\format{
An object of class \code{character} of length 24.
}
\usage{
BB_colors
}
\description{
Custom palette for ggplot2 graphs
}
\keyword{datasets}
