% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_palettes.R
\name{scale_color_BB}
\alias{scale_color_BB}
\title{Function to use custom colors in ggplot2}
\usage{
scale_color_BB(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Function to use custom colors in ggplot2
}
